+++ tests/torch/test_graph_building.py
@@ -14,7 +14,6 @@ import pytest
 import torch
 import torch.nn.functional as F
 from copy import deepcopy
-from pkg_resources import parse_version
 from torch import nn
 from typing import List
 from typing import Tuple
@@ -585,61 +584,3 @@ def test_integer_path_marking():
     assert num_integer_edges == 4
 
 
-@pytest.mark.skipif(parse_version(torch.__version__) < parse_version("1.11"),
-                    reason="__getitem__ works unexpectedly for TracedTensor until fix in torch 1.11.\n"
-                           "Fix in pytorch: https://github.com/pytorch/pytorch/pull/67202\n"
-                           "Related ticket: 82065")
-def test_torch_tensor_getitem_behavior(mocker):
-    x = torch.ones((10, 4, 4, 4))
-    indexes = torch.LongTensor([0, 1, 2])
-    mock_tensor_meta = mocker.stub
-    from nncf.torch.dynamic_graph.trace_tensor import TracedTensor
-    traced_x = TracedTensor.from_torch_tensor(torch.ones((10, 4, 4, 4)), mock_tensor_meta)
-    traced_indexes = TracedTensor.from_torch_tensor(torch.LongTensor([0, 1, 2]), mock_tensor_meta)
-    SHAPE_1 = [3, 4, 4, 4]
-    SHAPE_2 = [3, 4, 4]
-    assert list(x[indexes].shape) == SHAPE_1
-    assert list(x[traced_indexes].shape) == SHAPE_1
-    assert list(traced_x[indexes].shape) == SHAPE_1
-    assert list(traced_x[traced_indexes].shape) == SHAPE_1
-
-    assert list(x[indexes, indexes].shape) == SHAPE_2
-    assert list(x[traced_indexes, traced_indexes].shape) == SHAPE_2
-    assert list(traced_x[indexes, indexes].shape) == SHAPE_2
-    assert list(traced_x[traced_indexes, traced_indexes].shape)
-    assert list(x[indexes, traced_indexes].shape) == SHAPE_2
-    assert list(traced_x[indexes, traced_indexes].shape) == SHAPE_2
-
-
-class ModelSpecificException(Exception):
-    pass
-
-
-class ExceptionRaisingModule(torch.nn.Module):
-    class Inner(torch.nn.Module):
-        def __init__(self):
-            super().__init__()
-            self.dummy_param = torch.nn.Parameter(torch.ones([1]))
-
-        def forward(self, *args, **kwargs):
-            _ = torch.cat((torch.ones([1]), torch.ones([1])))
-            raise ModelSpecificException
-
-    def __init__(self):
-        super().__init__()
-        self.seq = torch.nn.Sequential(self.Inner(), self.Inner())
-
-    def forward(self, *args, **kwargs):
-        return self.seq(*args, **kwargs)
-
-
-def test_scope_and_call_counters_are_reset_on_exceptions():
-    ctx = TracingContext()
-    model = ExceptionRaisingModule()
-    with pytest.raises(ModelSpecificException):
-        with ctx:
-            model(torch.ones([1]))
-    assert not ctx.module_call_stack
-    assert not ctx.relative_scopes_stack
-    #pylint:disable=protected-access
-    assert not ctx._threading.thread_local.operator_counters
